recursive_cte_demo/requirements.txt
Django==1.9.1
django-cors-headers==1.1.0
djangorestframework==3.3.3
psycopg2==2.6.1
pyexcel-io==0.2.1
pyexcel-xls==0.2.1
pylint==1.5.0
pymongo==3.2
pystache==0.5.4
python-dateutil==2.5.3
urllib3==1.15.1
wheel==0.24.0
xlrd==1.0.0
xlutils==2.0.0
xlwt==1.1.2
django-cte-trees==1.0b1
django-cors-headers==1.1.0